requirements.txt
Flask==3.0.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pymongo==4.6.2
groq==0.4.2
google-generativeai==0.4.0
ultralytics==8.1.0
opencv-python==4.9.0.80
requests==2.31.0
